Chore: Restructure files for customReasons components
