+++ react/admin/settings/components/CustomReasons/index.ts
@@ -1,0 +1,1 @@
+export { CustomReasons } from './CustomReasons'
